pyproject.toml
[tool.poetry]
name = "skp2-drf-online-courses"
version = "0.1.0"
description = ""
authors = ["Stereo2025 <teneloks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
stripe = "^9.4.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
